rent-agent/pyproject.toml
[project]
name = "rent-agent"
version = "0.1.0"
description = "LangGraph-based JoyRent agent"
requires-python = ">=3.10"
dependencies = [
    # ==================== LangGraph 核心 ====================
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    
    # ==================== LLM SDK ====================
    "openai>=1.0.0",  # 通义千问兼容 OpenAI SDK
    "dashscope>=1.20.0",  # 通义千问官方 SDK（备用）
    
    # ==================== 数据库驱动 ====================
    "psycopg2-binary>=2.9.0",  # PostgreSQL
    "pgvector>=0.3.0",  # pgvector 扩展
    "pymysql>=1.1.0",  # MySQL
    "sqlalchemy>=2.0.0",  # ORM
    
    # ==================== 向量检索 ====================
    "faiss-cpu>=1.8.0",  # 本地向量检索（可选）
    "numpy>=1.24.0",
    
    # ==================== 缓存 ====================
    "redis>=5.0.0",  # Redis 客户端
    
    # ==================== HTTP 请求 ====================
    "httpx>=0.27.0",  # RAWG API 调用
    "requests>=2.31.0",
    
    # ==================== 工具库 ====================
    "python-dotenv>=1.0.0",  # 环境变量加载
    "pydantic>=2.0.0",  # 数据验证
    "pydantic-settings>=2.0.0",
    
    # ==================== Web 框架（API 服务）====================
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # ==================== 日志和监控 ====================
    "loguru>=0.7.0",
    "prometheus-client>=0.21.0",  # 可观测性
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

